subsystems/cont_proc/requirements.txt
streamlit
pillow==8.4.0
matplotlib==3.4.3
numpy==1.19.5
pandas==1.4.1
pymongo==4.0.2
fitter==1.4.0
sklearn==0.0
seaborn
confluent-kafka==1.9.2
scipy
scikit-learn
uuid
fastavro==1.9.0
plotly==5.18.0